Refactored the WorklogQueryResource class.
